Divide WCA.py to each & change for python style
