+++ Cluster.py
@@ -1,0 +1,22 @@
+class _cluster(object):  # cluster(= set of nodes) class
+    def __init__(self):
+        self.nodes = []
+        self.feature_vector = []  # each cluster has own feature vector generated by nodes' feature vector(f1+f2/n1+n2 term)
+
+    def get_nodes(self):
+        return self.nodes[:]
+
+    def get_feature_vector(self):
+        return self.feature_vector[:]
+
+    def add_node(self, _input):
+        self.nodes.append(_input)
+
+    def set_feature_vector(self, _input):
+        self.feature_vector = _input[:]
+
+    def __copy__(self):
+        new = _cluster()
+        new.nodes = self.nodes[:]
+        new.feature_vector = self.feature_vector[:]
+        return new
